server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
python-multipart
# AI & Cache
opena
pdf2image
pydub